Renaming ATLAS images to twaek the sort.
We want a nice picture loaded as default when opening the ATLAS page. 
The deafult one is the obne whose thumbnail gets placed under the cursor 
at the center of the screen. I noticed that the 5th image got loaded as 
the default one; so, I swapped the 2nd with the 5th image. Now the 2nd 
is loaded when the user opens the ATLAS page.